backend/requirements.txt
asgiref==3.5.2
blessed==1.19.1
bpython==0.23
certifi==2022.9.24
charset-normalizer==2.1.1
curtsies==0.4.0
cwcwidth==0.1.7
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
django-dirtyfields==1.9.2
greenlet==1.1.3
idna==3.4
Pygments==2.13.0
PyJWT==2.5.0
pytz==2022.4
pyxdg==0.28
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
wcwidth==0.2.5
uwsgi==2.0.27
psycopg2-binary
fuzzywuzzy==0.18.0
pydantic==2.4.2
python-Levenshtein==0.23.0
django-constance==4.1.2
jinja2==3.1.4
boto3==1.35.34
aws-secretsmanager-caching==1.1.3
ec2_metadata==2.13.0
wheel==0.37.1
setuptools==68.2.2
ipython
